CHROMIUM: drm/panel: lpm102a188a: Disable backlight immediately
With the display dimming enabled, the panel takes a little while to
ramp the LED PWM down to zero. In the case where the panel should be
disabled or dark, we don't want to rely on this asynchronous dimming
since artifacts may become visible on the screen.

To fix this, disable the backlight controller entirely when a NULL
brightness is desired.

BUG=chrome-os-partner:45829
TEST=Tested on smaug with build 86. 100 power button cycles, no flash

Change-Id: Ide1a7e14be619f65110a3ab6566b90b279ce7be9
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/302600
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Sean Paul <seanpaul@google.com>