FROMLIST: genirq: Allow migration of chained interrupts by installing default action
When a CPU is offlined all interrupts that have an action are migrated to
other still online CPUs. However, if the interrupt has chained handler
installed this is not done. Chained handlers are used by GPIO drivers which
support interrupts, for instance.

When the affinity is not corrected properly we end up in situation where
most interrupts are not arriving to the online CPUs anymore. For example on
Intel Braswell system which has SD-card card detection signal connected to
a GPIO the IO-APIC routing entries look like below after CPU1 is offlined:

  pin30, enabled , level, low , V(52), IRR(0), S(0), logical , D(03), M(1)
  pin31, enabled , level, low , V(42), IRR(0), S(0), logical , D(03), M(1)
  pin32, enabled , level, low , V(62), IRR(0), S(0), logical , D(03), M(1)
  pin5b, enabled , level, low , V(72), IRR(0), S(0), logical , D(03), M(1)

The problem here is that the destination mask still contains both CPUs even
if CPU1 is already offline. This means that the IO-APIC still routes
interrupts to the other CPU as well.

We solve the problem by providing a default action for chained interrupts.
This action allows the migration code to correct affinity (as it finds
desc->action != NULL).

Also make the default action handler to emit a warning if for some reason a
chained handler ends up calling it.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
(am from https://patchwork.kernel.org/patch/7325891)
Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>

BUG=chrome-os-partner:45628
TEST=`echo 0 > /sys/devices/system/cpu/cpu1/online`
     Verify input devices continue to respond to user
     input.

Change-Id: I1f3f3d4eb3dd93e926234739070022dd95dfe149
Reviewed-on: https://chromium-review.googlesource.com/301592
Commit-Ready: Ullysses A Eoff <ullysses.a.eoff@intel.com>
Tested-by: Russ Sage <russ.sage@intel.com>
Reviewed-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Ganesh R Patil <ganesh.r.patil@intel.com>