UPSTREAM: PM / sleep: Add flags to indicate platform firmware involvement
There are quite a few cases in which device drivers, bus types or
even the PM core itself may benefit from knowing whether or not
the platform firmware will be involved in the upcoming system power
transition (during system suspend) or whether or not it was involved
in it (during system resume).

For this reason, introduce global system suspend flags that can be
used by the platform code to expose that information for the benefit
of the other parts of the kernel and make the ACPI core set them
as appropriate.

Users of the new flags will be added later.

BUG=chrome-os-partner:43079
TEST=None

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit ef25ba0476015908ef5960f9faac149ddf34ede0)
Signed-off-by: Derek Basehore <dbasehore@chromium.org>

Change-Id: I95fcd1b0aa118a062c4e020d8361c2d576a1ae1a
Reviewed-on: https://chromium-review.googlesource.com/308178
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>