UPSTREAM: input: i8042: Avoid resetting controller on system suspend/resume
If the upcoming system suspend is not going to be handled by the
platform firmware, like in the suspend-to-idle case, it is not
necessary to reset the controller in i8042_pm_suspend(), so avoid
doing that.

Moreover, if the system resume currently in progress has not been
started by the platform firmware, like in the suspend-to-idle case,
i8042_controller_resume() need not be called by i8042_pm_resume(),
so avoid doing that too in that case.

Additionally, try to catch the event that woke up the system by
calling the interrupt handler early during system resume if it has
not been started by the platform firmware.

BUG=chrome-os-partner:43079
TEST=on glados
echo freeze > /sys/power/state
wait 5 seconds
press keyboard key and check that system wakes up

Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Acked-by: Dmitry Torokhov <dmitry.torokhov@gmail.com>
(cherry picked from commit 1c5dd13459fe7caaeda82f7be878720c978d26f2)
Signed-off-by: Derek Basehore <dbasehore@chromium.org>

Change-Id: I46c319b78bac11b2fd08733733b91de59dbb1058
Reviewed-on: https://chromium-review.googlesource.com/308263
Commit-Ready: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>