CHROMIUM: host1x: Implement per-user clock constraints
Implement functions to hand over control of tegradrm client clocks to
host1x, so that host1x can manage the aggregation of per-user clock
constraints in a unified manner.

When a channel is opened, we add a struct host1x_user to the
struct host1x_client (and this struct host1x_user is removed on
close_channel).  Whenever a user wants to set a clock (via an ioctl
that is not part of this patchset), instead of setting the clock directly,
we use the host1x client's user list to aggregate all of the constraints
from other users to come up with a final clock setting.

We also provide different ways of setting the clock.  We can directly ask
for a particular clock rate, or (in the case of memory bandwidth controlled
by the emc clocks), we can ask for bandwidth.

Since host1x is aware of some tegradrm client clocks, implement some helper
functions to enable and disable these clocks, and switch the falcon-based
tegradrm clients over to using them.

In the tegradrm clients' probe() and remove() methods, since we are letting
host1x manage some clocks, we need to register the host1x client before
calling power on.

BUG=chrome-os-partner:42645
TEST=Build and boot on smaug.  Play video on youtube, and take picture with
camera app, to make sure falcon-based tegradrm clients still probe and work.

Change-Id: I65bf6610d885cd6ed09bc4f7bada1da6ed830bf1
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/305062
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>