CHROMIUM: host1x: Add host1x_user to channel get/put
That way, we can start to track individual users of a host1x device,
with the assumption that a channel open/close session defines a user
session.

Right now, struct host1x_user is an empty placeholder.  We will need to
track users to, for example, implement per-user clock constraints.

BUG=chrome-os-partner:42645
TEST=Build and boot on smaug.

Change-Id: Iff8f9435d6b352888f539d14b16b636647db014b
Signed-off-by: Andrew Chew <achew@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/305061
Reviewed-by: Arto Merilainen <amerilainen@nvidia.com>
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>