CHROMIUM: mfd: max77620: Update max77620_regmap_config_unlock
The max77620_regmap_config_lock|unlock routines have identical
checks on shutdown, in_atomic, and irqs_disabled to determine if
the lock should be used or not.

However, it seems that in rare cases, those conditions can change
in between calling *_lock and *_unlock. This can cause situations
of unbalanced locking, either trying to unlock a lock that was
never taken, or skipping unlocking a lock that was.

To ensure this never happens, use a boolean to store whether or
not the lock was actually taken, and simply use that to determine
whether or not to unlock.

BUG=None
TEST=shutdown system repeatedly and verify no unbalanced unlocks occur

Change-Id: I72984749aaace2fa5e78c6b385a3ff716079c496
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/304929
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
(cherry picked from commit e1722e8c1fa2002d8aa0dec3da8bab05ff5ccd7e)
Reviewed-on: https://chromium-review.googlesource.com/307251
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>