CHROMIUM: jdi-lpm102a188a: fix unbalanced regulator disable on shutdown
Recent regulator work in this driver caused a regression in the
shutdown path that would disable already disabled regulators.  This
causes the "unbalanced disables for PPLCD_VDD" messages on those regulators.

Properly unprepare the panel after we disable it.  This effectively checks
the enabled flag to see if we need to turn off the regulators.

Signed-off-by: Bryan Freed <bfreed@chromium.org>

BUG=chrome-os-partner:46319
TEST=check for console errors on 'reboot' and (via UI) PowerOff

Change-Id: I894179245f84e8a921eab4c4421576a48f611bb0
Reviewed-on: https://chromium-review.googlesource.com/304876
Commit-Ready: Bryan Freed <bfreed@google.com>
Tested-by: Bryan Freed <bfreed@google.com>
Reviewed-by: Benson Leung <bleung@chromium.org>