CHROMIUM: drm/nouveau: Don't free fence context until fence is released
In commit 6939d67224 ("CHROMIUM: drm/noveau: Fix fence context leak"),
fence context reference counting was removed so that the fence context
was freed directly in the nv*_fence_context_del() path.  It's possible,
however, that references to the fence may still exist.  Since the fence
context contains the fence's lock, the fence context must not be freed
until the last reference to the fence is dropped.  To prevent the
potential use after free here, restore the reference counting removed by
the above commit and add a release() method to nouveau_fence_ops_uevent
so that we don't end up leaking the fence context.

BUG=chrome-os-partner:46382
TEST=Reboot cycle test on Smaug; no crashes seen.

Change-Id: Id58750f4d3d54e0629dc201391bd9e881fb5cb33
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305330
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Reviewed-by: Xia Yang <xiay@nvidia.com>