Revert "CHROMIUM: ARM64: tegra: smaug: Enable touchscreen reset."
This reverts commit c84047dac3ba59e22a1f5ea8c390f02517f6a458.

Recent touchscreen firmware changes have made it possible to save nearly
as much power in S3 as when we forcibly reset the device and turned off
its power supplies.  As those techniques occasionally introduced noise
on the sensor its best to leave it powered and running if power is
acceptable.

Signed-off-by: Todd Broch <tbroch@chromium.org>

BUG=chrome-os-partner:43346
TEST=measure power in S3 w/ new touchscreen fw and its nearly as low as
when its held in reset and power rails are turned off.

Change-Id: I0c63ae31c9fbe6789c6d85e688150d2f8641e5ef
Reviewed-on: https://chromium-review.googlesource.com/306613
Commit-Ready: Todd Broch <tbroch@chromium.org>
Tested-by: Benson Leung <bleung@chromium.org>
Tested-by: Todd Broch <tbroch@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>