Revert "Revert rt5677 hotwording support"
This is a reland of hotwording support on rt5677. In a previous
commit we changed that GPIO_CTRL1 and GPIO_CTRL2 registers can be
configured as volatile using dtsi config, which fixes the original
issue that speaker amp and dmic enable pin being reset across
suspend/resume.

This reverts commit 9abbf21f14b12242e12d503d455765ace39e87ae.

BUG=b:24334115
TEST=Update kernel to Smaug, verify hotwording works.

Change-Id: I23b859690509222ab92c09d37cf1bdbc582113e4
Signed-off-by: Hsin-Yu Chao <hychao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/307663
Commit-Ready: Hsinyu Chao <hychao@chromium.org>
Tested-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Anatol Pomazau <anatol@google.com>
Reviewed-by: Ben Zhang <benzh@chromium.org>