CHROMIUM: iio: cros_ec: Alter timestamp to handle IRQ latency.
It can take some time for the AP to process IRQ from the EC: the EC may
have added extra sensor samples, but the not the final timestamp.
Use the timestamp noted when retrieving EC FIFO information as the final
timestamp if the final timestamp is not present in the FIFO.

BUG=chrome-os-partner:43811
TEST=Check that SensorIntegrationTests now passes persistently.
Check with dmesg output that the changes is indeed needed often.

Change-Id: I6b9dece9614bf3f78e762200b87a932d2e47f969
Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/307824
(cherry picked from commit 1192824ba811612f6e17ed3bab7b38bfb77a28d2)
Reviewed-on: https://chromium-review.googlesource.com/309570