UPSTREAM: ASoC: Intel: Skylake: Add support for virtual dsp widgets
In SKL topology routes, some paths can be connected by a widget
which are not a DSP FW widget and virtual with respect to
firmware. In these case when module has to bind, then the
virtual DSP modules needs to skipped till a actual DSP module is
found which connects the pipelines.

So we need to walk the graph and find a widget which is real in
nature. This patch adds that support and splits
skl_tplg_pga_dapm_pre_pmu_event() fn with parsing code to
skl_tplg_bind_sinks() fn and call that recursively as well as
while parsing

The patch moves code a bit while splitting so diffstat doesn't
tell real picture

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
br: for-next
(cherry picked from commit 8724ff17521a91a87971027cf78631030091bc52)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:41280
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: I03551f01ee8dc0bb64f1fc9ef238000d14b6a852
Reviewed-on: https://chromium-review.googlesource.com/310374
Commit-Ready: James Ausmus <james.ausmus@intel.com>
Tested-by: Vedang Patel <vedang.patel@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>