CHROMIUM: net: wireless: bcmdhd: no retry on rxglom superframe error
Aborting the current read attempt on the superframe also removes the packet
from the pipeline. Retries should not be attempted on the next packet since
it would not be a superframe(either a superframe descriptor or other data
packet) and should not be handled by dhdsdio_rxglom.

BUG=chrome-os-partner:45478
TEST=manual test with iperf on ryu

Change-Id: I77fbb64d30ca1951aeb47d24a09778de917d08fd
Signed-off-by: Franky Lin <frankyl@broadcom.com>
Reviewed-on: https://chromium-review.googlesource.com/310031
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>