CHROMIUM: net: wireless: bcmdhd: Fix argument to wl_cfg80211_netdev_notifier_call
The void * argument to wl_cfg80211_netdev_notifier_call() is a struct
netdev_notifier_info, not a struct net_device.  This results in
wl_dealloc_netinfo() being called with a pointer that cannot possibly
exist  in the net_device list (cfg->net_list).  wl_dealloc_netinfo()
silently  fails to delete the unregistered net_device from cfg->net_list;
subsequently the stale pointer to the freed net_device is used, resulting
in a kernel crash.

Fix this by passing the correct pointer to wl_dealloc_netinfo(), and print
a warning if we call wl_dealloc_netinfo() but the unregistered interface
is  not found in the list.

BUG=chrome-os-partner:45703

Change-Id: I190ec1312d2045dc83a15d3983f0d57a417f6a0d
Signed-off-by: Chilam Ng <chilam@broadcom.com>
Reviewed-on: https://chromium-review.googlesource.com/309973
Commit-Ready: Abhishek Bhardwaj <abhishekbh@google.com>
Tested-by: Abhishek Bhardwaj <abhishekbh@google.com>
Reviewed-by: Abhishek Bhardwaj <abhishekbh@google.com>