CHROMIUM: ARM: Add alt-syscall support
Add ARM support for changing a thread's syscall table at runtime using
the alt-syscall infrastructure.

Note that alt-syscall is not supported in kernel configurations using
OABI or EABI with OABI_COMPAT.  It's possible to support OABI, but it
requires a bit of re-work to the common alt-syscall infrastructure and
to the syscall entry path.  This should not be an issue since OABI is
not supported on ChromeOS or Android.

BUG=b:25408860
TEST=Build and boot on Minnie; syscalls still work with ALT_SYSCALL on.

Change-Id: Ie3373fe94b8398d88652e37d4bab45e921c82ee8
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/310972
Reviewed-by: Kees Cook <keescook@chromium.org>