CHROMIUM: drm/img-rogue/mt8173: Get all device resources during MTKMFGBaseInit
Current device tree binding lump all gpu resources together in a single
"mfgsys-gpu" node.  So, let's be consistent with this model and get all
device resources during MTKMFGBaseInit; both those for the 'mfg' device,
and those the mt8173 layer will pass for the 'RGX' gpu device in
img-rogue.

Resources (reg start/size & irq) for the "RGX" are cached in the mtk_mfg
struct, which is then used to 'fill in' the appropriate fields of the
PVRSRV_DEVICE_CONFIG passed as "SysConfig" data to the img-rogue core.

Note: Perhaps someday we can split the mfgsys-gpu node into two, one for
the "RGX" part, and the other for "mfg", and each could manage its own set
of resources directly.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:47249
TEST=build & sanity test webgl on oak-rev4

Change-Id: Ib844ddf6db851d46e3a2d7ecfee09c2fa2754c68
Reviewed-on: https://chromium-review.googlesource.com/311137
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Chiawen Lee <chiawen.lee@mediatek.com>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Rufus Hamade <rufus.hamade@imgtec.com>