CHROMIUM: drm/nouveau: Synchronize PRAMIN writes in map_pgt
Currently PRAMIN writes are just triggered by writes to PRAMIN window,
however, due to write posting, the actual write might be delayed. However,
we expect that page table mapping operation is visible to the GPU
immediately.

This patch makes the map_pgt operation perform a read-back of the just
written PDE to ensure that posted PRAMIN writes are complete.

BUG=chrome-os-partner:45900
TEST=More than 1000 reboots passed (together with one more CL)

Change-Id: I5b314848a7e297b83673417b541095a6e68c8145
Signed-off-by: Tomasz Figa <tfiga@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/305189
Tested-by: Puneet Kumar <puneetster@chromium.org>
Reviewed-by: Terje Bergstrom <tbergstrom@nvidia.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>
Reviewed-by: Alexandre Courbot <gnurou@gmail.com>