CHROMIUM: drm/tegra: dc: update LA/PTSA registers if plane's emc bandwidth is changed
Sometimes some plane's emc bandwidth increases, but the calculated dc's
emc bandwidth does not increase. It caused the bits for some plane in the
LA/PTSA registers not to be updated in time and may generate underflow

This patch adds the logic for updating the corresponding bits in the
LA/PTSA registers whenever any plane's emc bandwidth changes or the dc's
whole emc bandwidth changes. And this update follows the rules:
1) Update before any dc commit if any plane's emc bandwidth is
increasing, or none of plane's emc bandwidth changes but the dc's whole
emc bandwidth is increasing
2) Update after all dc commit if any plane's emc bandwidth is
decreasing, or none of plane's emc bandwidth changes but the dc's whole
emc bandwidth is decreasing

BUG=chrome-os-partner:47206
TEST=Add prints and verify that the logic works as expected

Change-Id: I6809c733debd0ef90205f9f15c42573e0a9e9b97
Signed-off-by: Kary Jin <karyj@nvidia.com>
[seanpaul cleaned up some of the code, no functional changes]
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/312402
Reviewed-by: Sean Paul <seanpaul@google.com>
Commit-Queue: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>