UPSTREAM: ASoC: Intel: Skylake: Fix CLDMA buffer wrap case
When downloading the firmware/module, if the ring buffer boundary
is reached, we need to wrap to the zeroth position. On next copy
we need to copy till end of buffer and the remaining buffer needs
to be copied from zeroth position.

In this case copy was not handled correctly when wrap condition
is reached which caused invalid data to be copied resulting in
invalid hash failure.

This patch fixes the issue by handling copy at the boundary
condition correctly.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit e797af53b8814dfbc3c6ac134c528b8ab480f275)
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
br: for-next
Signed-off-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>

BUG=chrome-os-partner:41280
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: I7c41c2f3356189d356c661d0bfa511bfa5c0ffcc
Reviewed-on: https://chromium-review.googlesource.com/312256
Commit-Ready: James Ausmus <james.ausmus@intel.com>
Tested-by: Vedang Patel <vedang.patel@intel.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>