HACK: mmc: sdhci: Retune SDIO device on resume
SDIO cards which keep power across suspend/resume do not undergo the
full re-initialization sequence on resume and thus are not retuned.
In this case, retuning should not be necessary and simply restoring the
tuned tap delay value obtained at boot should be sufficient.  This is
not working for the bcm4354 on Smaug, however, resulting in numerous CRC
errors post-resume.  To work around this, forcibly retune the controller
on resume if the card remained powered across suspend/resume.

Note that this issue does not exist in chromeos-3.18 since all
cards (regardless of whether or not they were powered in suspend) are
retuned on resume.

BUG=chrome-os-partner:45478
TEST=Suspend/resume on Smaug.

Change-Id: Id39ad953ae14ac6489942f81891effcd7e6ec5e0
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/312874
Reviewed-by: Kevin Cernekee <cernekee@chromium.org>
Reviewed-by: Christopher Freeman <cfreeman@nvidia.com>