CHROMIUM: drm/tegra: Fix wrong plane->crtc dereference
There is a kernel NULL pointer crash after this patch is merged:
"4de84b63c1d8 UPSTREAM: drm/atomic: fix out of bounds read in for_each_*_in_state helpers"

The reason for that is, in tegra drm driver, we deference the wrong
"plane->crtc". The crtc in an atomic commit is in "plane_state->crtc".
So this patch fixes this crash.

The reason why we didn't hit the crash before is, before apply that
upstream patch, in tegra_dc_evaluate_bandwidth, we have:

if (!crtc_state)
	break;

And this avoids calling tegra_dc_calculate_bandwidth which triggers
the crash. The upstream patch fixes the issue which crtc_state in
a "for_each_crtc_in_state" can be NULL(an out-of-band read) so
this further triggers the crash in tegra drm driver.

Change-Id: I77993ef30275bf84567ee0a9f04298b5fac94e04
Signed-off-by: Mark Zhang <markz@nvidia.com>
Reviewed-on: https://chromium-review.googlesource.com/313093
Commit-Ready: Sean Paul <seanpaul@google.com>
Tested-by: Sean Paul <seanpaul@google.com>
Reviewed-by: Sean Paul <seanpaul@google.com>
Reviewed-on: https://chromium-review.googlesource.com/313670
Reviewed-by: Andrew Bresticker <abrestic@chromium.org>
Commit-Queue: Andrew Bresticker <abrestic@chromium.org>
Tested-by: Andrew Bresticker <abrestic@chromium.org>